Move Resource to kernel
