+++ core/shared/src/main/scala/cats/effect/package.scala
@@ -68,6 +68,9 @@ package object effect {
   type ParallelF[F[_], A] = cekernel.Par.ParallelF[F, A]
   val ParallelF = cekernel.Par.ParallelF
 
+  type Resource[+F[_], +A] = cekernel.Resource[F, A]
+  val Resource = cekernel.Resource
+
   type OutcomeIO[A] = Outcome[IO, Throwable, A]
   type FiberIO[A] = Fiber[IO, Throwable, A]
 }
