+++ core/shared/src/test/scala/cats/effect/ResourceSpec.scala
@@ -17,7 +17,7 @@
 package cats.effect
 
 import cats.{~>, SemigroupK}
-import cats.data.{Kleisli} //, OptionT}
+import cats.data.Kleisli
 import cats.effect.kernel.Deferred
 import cats.effect.testkit.TestContext
 import cats.kernel.laws.discipline.MonoidTests
