+++ kernel/js/src/main/scala/cats/effect/kernel/ResourcePlatform.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package cats.effect
+package cats.effect.kernel
 
 /**
  * JS-specific Resource methods
