+++ kernel/shared/src/main/scala/cats/effect/kernel/Resource.scala
@@ -14,16 +14,15 @@
  * limitations under the License.
  */
 
-package cats.effect
+package cats.effect.kernel
 
 import cats._
 import cats.data.AndThen
 import cats.implicits._
-import cats.effect.implicits._
+import cats.effect.kernel.implicits._
 
 import scala.annotation.tailrec
 import Resource.ExitCase
-import cats.effect.kernel.Ref
 
 /**
  * The `Resource` is a data structure that captures the effectful
@@ -623,14 +622,6 @@ abstract private[effect] class ResourceInstances extends ResourceInstances0 {
       def F = F0
     }
 
-  implicit def catsEffectLiftIOForResource[F[_]](
-      implicit F00: LiftIO[F],
-      F10: Applicative[F]): LiftIO[Resource[F, *]] =
-    new ResourceLiftIO[F] {
-      def F0 = F00
-      def F1 = F10
-    }
-
   implicit def catsEffectCommutativeApplicativeForResourcePar[F[_]](
       implicit F: Async[F]
   ): CommutativeApplicative[Resource.Par[F, *]] =
@@ -779,14 +770,6 @@ abstract private[effect] class ResourceSemigroupK[F[_]] extends SemigroupK[Resou
 
       K.combineK(allocate(ra), allocate(rb))
     }
-}
-
-abstract private[effect] class ResourceLiftIO[F[_]] extends LiftIO[Resource[F, *]] {
-  implicit protected def F0: LiftIO[F]
-  implicit protected def F1: Applicative[F]
-
-  def liftIO[A](ioa: IO[A]): Resource[F, A] =
-    Resource.liftF(F0.liftIO(ioa))
 }
 
 abstract private[effect] class ResourceParCommutativeApplicative[F[_]]
